Use class names to subscribe events
Before:
  es.subscribe(Handler, ['OrderCreated', 'OrderExpired'])

After:
  es.subscribe(Handler, [OrderCreated, OrderExpired])

This commit makes subscriptions consistent. Previously
handlers were supplied as a class and events as a string.
After this change strings are no longer used.

Other than that by using class names you can ensure
that the subscribed event really exists. If to doesn't
then a NameError is raised.

In order to achieve that the ability to specify an event_type
in constructor has been removed.

    TestEvent.new(event_type: 'OrderCreated') # doesn't work anymore

Of course this change is not backward compatible.
I think it would be possible to introduce a compatibility layer
and for the time being allow to subscribe by both class names and
strings.